Refactor the handleClic in the App, delete debugging code and move calculations operations in calculate function file
